pulled out expand_attributes module
